Renamed scripts to match new version of course
