+++ Assets/Scripts/CompletedHelloWorld.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class HelloWorld : MonoBehaviour
+public class CompletedHelloWorld : MonoBehaviour
 {
     // Start is called before the first frame update
     void Start()
